Implement LLDPA module
Summary:
For TX_REQ, it sends out LLDP PDU for a specified interval
For RX_REQ, it expects to receive incoming LLPD PDU for a
specified interval. It sends Timeout Msg to controller if
it doesn't receive LLDPDU within that interval.
Switch forwards LLDPDU as packet-in to the controller if
LLDPDU is unexpected (port not set or data not matched)